# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Daycore v2 API
  version: "2.4.0"
  description: >
    Daycore v2 backend (Go). Autonomous AI day-planning + mood companion.
    Anonymous-first: POST /api/session/init issues an httpOnly, signed session
    cookie (dc_sid) that authorizes all data endpoints. Optional email/OAuth
    sign-in sets an httpOnly JWT cookie (dc_auth); logout revokes all issued JWTs
    (server-side token version). Same-origin web frontends MUST send cookies
    (fetch: credentials:"include"). Native apps and separately-deployed frontends
    use header credentials instead (dual-rail, cookies stay supported): the
    signed session token in X-Session-Token (returned by /api/session/init with
    {"tokenInBody":true}) and/or the JWT in Authorization: Bearer (returned by
    login/register with {"tokenInBody":true}). Check GET /api/version (apiVersion
    contract major) before use — matching rules in FRONTEND_HANDOFF.md.
    The browser extension authenticates to /api/import/* with X-Import-Token;
    admin endpoints use X-Admin-Token.
    Errors are {error, message}. The single SSE endpoint is /api/ai/companion —
    see FRONTEND_HANDOFF.md §B for the v2 frame protocol (NOT the legacy [DONE]).
servers:
  - url: http://localhost:8080
    description: Local (Go serves the API; nginx proxies /api in production)
...
security:
  - sessionCookie: []
  - sessionToken: []
  - bearerAuth: []
paths:
...
  /api/v2/plan:
    get:
      operationId: getPlan
      tags: [plan]
      summary: Single day with rule occurrences merged in (JSON null when empty).
      parameters: [{ name: date, in: query, required: true, schema: { type: string, example: "2026-07-13" } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { oneOf: [ { $ref: '#/components/schemas/DayPlan' }, { type: 'null' } ] } } } } }
    post:
      operationId: upsertPlan
      tags: [plan]
      summary: Upsert a day's plan (full blocks). Use PATCH for single-block edits.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [date], properties: { date: { type: string }, blocks: { type: array, items: { $ref: '#/components/schemas/TimeBlock' } }, note: { type: string }, sourceType: { type: string } } } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DayPlan' } } } } }
    patch:
      operationId: patchPlan
      tags: [plan]
      summary: One incremental block action (add/update/remove); materializes rule blocks + tombstones.
      description: >
        `date` is a routing key, not a mutable block field: plans are sharded by
        day, so `changes.date` is ignored. A cross-day move is remove+add (or a
        reschedule), never an update of the block's date in place.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [date, action], properties: { date: { type: string }, action: { $ref: '#/components/schemas/PlanAction' } } } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DayPlan' } } } }
        '409':
          description: The plan gate refused — the block is locked or has petrified.
          content: { application/json: { schema: { $ref: '#/components/schemas/PlanBlocked' } } }
  /api/v2/plan/lock:
    post:
      operationId: lockPlanBlock
      tags: [plan]
      summary: Pin a block's time by hand, or let it go — the way out of a lock refusal.
      description: >-
        Sets lockLevel with lockSource=user, which is what stops derivation from
        putting the old lock back on the next write. Unlocking clears the reason:
        "set by your class timetable" should not stay attached to something the
        timetable no longer governs. Resetting to "derived" is deliberately not
        offered — see docs/specs/plan-semantics.md.
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [date, blockId, level], properties: {
          date: { type: string },
          blockId: { type: string },
          level: { type: string, enum: [none, soft, hard] },
          reason: { type: string, description: "Optional, ≤200 runes. Kept verbatim and never translated — they are the user's own words. Empty renders as the default reason for the level, in the reader's language." } } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DayPlan' } } } }
        '404': { description: No such block on that day. }
        '409':
          description: The plan gate refused — the block has petrified, so its lock is history too.
          content: { application/json: { schema: { $ref: '#/components/schemas/PlanBlocked' } } }
  /api/v2/plan/range:
    get:
      operationId: getPlanRange
      tags: [plan]
      summary: Batch fetch [from,to] with rule occurrences merged per date.
      parameters: [{ name: from, in: query, required: true, schema: { type: string } }, { name: to, in: query, required: true, schema: { type: string } }]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/DayPlan' } } } } } }
  /api/v2/plan/reschedule:
    post:
      operationId: proposeReschedule
      tags: [plan]
      summary: Turn "重新安排" into a proposal ghost, not an immediate edit.
      description: >-
        提案永远是虚影: rescheduling a petrified block is a suggestion the user
        nods to, never a write the server performs on its own. Returns a pending
        timed proposal at tomorrow's same slot; accepting it runs the refish chain.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [blockId], properties: { blockId: { type: string } } } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Proposal' } } } }
        '404': { description: no such block }

  # ─── proposals ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: dc_sid }
    authCookie: { type: apiKey, in: cookie, name: dc_auth }
    sessionToken: { type: apiKey, in: header, name: X-Session-Token, description: "Same signed value as the dc_sid cookie, for clients without a cookie jar. Get it from POST /api/session/init {tokenInBody:true}." }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: "Same JWT as the dc_auth cookie. Get it from login/register {tokenInBody:true}; logout revokes all issued JWTs (token-version bump)." }
    importToken: { type: apiKey, in: header, name: X-Import-Token, description: "Session-scoped token for the browser extension's direct push." }
    adminToken: { type: apiKey, in: header, name: X-Admin-Token, description: "Admin endpoints." }
...
  schemas:
...
    TimeBlock:
      type: object
      required: [id, title, type, time_mode, timezone]
      properties:
        id: { type: string }
        date: { type: string, example: "2026-07-13", description: "YYYY-MM-DD" }
        time: { type: string, nullable: true, example: "14:30", description: "HH:MM or null when unscheduled" }
        title: { type: string }
        type: { type: string, enum: [task, appointment, break, relax, meal] }
        duration_min: { type: integer, nullable: true }
        time_mode: { type: string, enum: [floating, fixed, local], description: "floating=wall clock (no utc_time); fixed=absolute instant; local=circadian, follows current tz. fixed/local carry utc_time." }
        timezone:
          type: string
          example: "Asia/Shanghai"
          description: >
            The EVENT's zone — whose wall clocks this rule's time belongs to.
            A different question from Preferences.timezone, which is where the
            user is now: a lecture imported from a Shanghai timetable stays
            09:00 中国时间 when the student is in London, and the two layers are
            what make that expressible. Empty means floating (whatever the
            reader's own zone says), which is right for "gym at 7" and wrong for
            a class in another country. Set by ICS import from X-WR-TIMEZONE and
            correctable here; a name this build cannot load is rejected.
        utc_time: { type: string, nullable: true, description: "RFC3339 UTC anchor for fixed/local blocks." }
        offset_min: { type: integer, nullable: true }
        offset_ref: { type: string }
        completed: { type: boolean }
        isAchievement: { type: boolean }
        rule_id: { type: string, description: "Set when the block is an occurrence of a ScheduleRule." }
        origin: { type: string, enum: [auto, manual, rule], description: "auto=generated (dashed border, replaced on re-plan); manual=user (survives); rule=materialized occurrence." }
        hidden: { type: boolean, description: "Tombstone — filtered out of reads; never render." }
        note: { type: string, description: "Whatever the user felt like writing when checking the block off — how it went, what was left half-done. Always optional; never prompted for. The agent reads it." }
        lock_level:
          type: string
          enum: ["", none, soft, hard]
          description: >-
            How firmly the time slot is fixed. THREE states, not two: "" means
            never derived yet (the server will fill it in on read), none means
            derived and free to move. hard=set by a timetable, the user cannot
            move it; soft=agreed with someone else, moving needs confirmation.
            Derivation rule and the shared case table live in api/lock-rules.json.
        lock_reason: { type: string, description: "Why it is locked. Derived reasons are looked up per request locale; a user-set reason is stored verbatim." }
        lock_source:
          type: string
          enum: ["", derived, user, agent]
          description: >-
            Who decided the lock. derived=inferred from type+title and safe to
            recompute; user/agent=a deliberate choice, never recomputed. Without
            this a hand-unlocked class would be re-locked on every read.
        rescheduled_from: { type: string, description: "Id of the block this one was re-fished from, chaining each retry back to the original." }
        reschedule_count: { type: integer, description: "How many times this item has been offered a new slot. Past the configured limit the agent stops offering and records the preference instead." }
    DayPlan:
      type: object
      properties:
        id: { type: string, description: "Empty for synthetic rules-only days." }
        sessionId: { type: string }
        date: { type: string }
        blocks: { type: array, items: { $ref: '#/components/schemas/TimeBlock' } }
        sourceType: { type: string, enum: [text, image, auto, rules] }
        note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PlanAction:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [add, update, remove] }
        match: { type: object, additionalProperties: true, description: "e.g. {id}" }
        changes: { type: object, additionalProperties: true }
        confirm:
          type: boolean
          description: >-
            The way through a soft lock: the user was told this one was agreed
            with someone else and said go ahead. Does nothing for a hard lock or
            a petrified block — those have no way through, and a flag that
            opened them would make the gate advisory. See docs/specs/plan-semantics.md.
    Proposal:
      type: object
      description: >-
        Everything the agent wants to do but has not done — or has done and is
        still offering to undo. Ghosts, cards, decision cards and channel
        approvals are all this, seen from different ends. Semantics in
        docs/specs/plan-semantics.md §three.
      properties:
        id: { type: string }
        state:
          type: string
          enum: [pending, accepted, rejected, expired]
          description: >-
            Four, not three. A silent expiry must stay distinguishable from a
            deliberate no — the ledger keeps everything, and an expired offer
            was never seen.
        level:
          type: string
          enum: [L1, L2, L3]
          description: >-
            The attention ladder — the MEDIUM. L1 an in-view ghost bound to a
            slot, L2 a card in the stack, L3 a push against the ≤3/day budget.
            L0 never becomes a proposal. Independent of whether the proposal may
            ESCALATE, which is decided by what it rests on (see origin/btype):
            a hard-backed one — an appointment, a deadline — may come back louder
            as the moment approaches; a soft one gets one delivery, ever, because
            "ignoring is always safe" is a promise and escalation would break it.
            The distinction is derived server-side from the row, never asserted
            by the agent: "is this important" is exactly the judgement a model
            gets wrong in the direction of interrupting more.
        kind: { type: string, enum: [timed, card, decision] }
        title: { type: string }
        summary: { type: string }
        reason: { type: string }
        evidence: { type: string }
        date: { type: string }
        start: { type: string, description: "HH:MM" }
        dur: { type: integer, nullable: true }
        btype:
          type: string
          description: >-
            The block type this proposal is about. "appointment" is the only
            HARD one today — a slot somebody else is also holding — and future
            hard types (meeting, flight, exam) join that set rather than adding a
            field. task/break/relax/meal are the user's own intentions about
            their own day. Deliberately NOT the same as lockLevel: a lock says
            "this cannot move", hardness says "missing this costs somebody other
            than you".
        lockLevel: { type: string }
        lockReason: { type: string }
        rows:
          type: array
          description: A compound card, each line accepted or rejected on its own.
          items:
            type: object
            properties:
              id: { type: string }
              label: { type: string }
              state: { type: string, enum: [pending, accepted, rejected, expired] }
        sessionId: { type: string }
        mergeKey: { type: string, description: Same key → only the newest undelivered survives. }
        deliverAfter: { type: string, format: date-time }
        deliveredAt:
          type: string
          format: date-time
          description: >-
            A PERMANENT stamp; nothing clears it. "How many cards are in the
            stack" is a conjunction (pending + delivered + still deliverable),
            not a question about this field.
        pushedAt: { type: string, format: date-time, description: What the ≤3/day push budget counts. }
        ttlPolicy:
          type: string
          enum: [silence_accepts, silence_rejects]
          description: >-
            Acted first → silence confirms; asked first → silence voids.
            Backwards, one of these executes unattended work and the other
            discards an undo window.
        expiresAt: { type: string, format: date-time }
        resolution: { type: string, enum: [user, silence, superseded] }
        origin:
          type: string
          enum: [agent, daemon, upload, protector, user, deadline]
          description: >-
            Which part of the system produced it. "deadline" is the fact track —
            a due date the user or an import put there, the archetypal hard fact,
            and the one origin that licenses escalation. "user" means the user
            asked for this card; it matters for the ledger, since a summoned card
            should not count against rapport the way an unsolicited one does.
        threadId: { type: string }
        rev: { type: integer, description: Optimistic concurrency — row-level accepts race otherwise. }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PlanBlocked:
      type: object
      description: >-
        409 body when the plan gate refuses a write. 409 rather than 403 because
        nothing about the caller is wrong — the request conflicts with the state
        of the block it names, and that state can change.
      required: [error, code, message, blockId, confirmable]
      properties:
        error: { type: string, enum: [blocked] }
        code:
          type: string
          enum: [locked, petrified]
          description: Which rule refused. `petrified` means the day has turned; `locked` means the slot is not the user's to move.
        message: { type: string, description: Localized, ready to show. }
        blockId: { type: string }
        confirmable:
          type: boolean
          description: Re-sending with `action.confirm` would succeed. True only for soft locks.
        lockLevel: { type: string, enum: [hard, soft], description: Absent for a petrify refusal. }
        lockReason: { type: string, description: Absent for a petrify refusal. }
        block: { $ref: '#/components/schemas/TimeBlock' }
...